css compacted alert/notice area when empty structured top/models/creators/collections into 2nd row menu: keeping filters active between the different "views" pulled filters into a concern to share code between the different views added tree-based sql queries for handling collection nesting display and searching
