+++ app/controllers/collections_controller.rb
@@ -1,15 +1,28 @@
 class CollectionsController < ApplicationController
+  include ModelFilters
   before_action :get_collection, except: [:index, :new, :create]
 
   def index
-    @collections =
-      if current_user.pagination_settings["collections"]
-        page = params[:page] || 1
-        Collection.all.page(page).per(current_user.pagination_settings["per_page"])
-      else
-        Collection.all
-      end
-    @title = "Collections"
+    process_filters_init
+    process_filters_tags_fetchall
+    process_filters
+    process_filters_tags_highlight
+
+    # @collections = Collection.where(id: @models.map{|model| model.collection_id})
+    @collections = Collection.tree_both(@filters[:collection]||nil,@models.map{|model| model.collection_id}.compact)
+
+    # Ordering
+    @collections = case session["order"]
+    when "recent"
+      @collections.order(created_at: :desc)
+    else
+      @collections.order(name: :asc)
+    end
+
+    if current_user.pagination_settings["collections"]
+      page = params[:page] || 1
+      @collections = @collections.page(page).per(current_user.pagination_settings["per_page"])
+    end
   end
 
   def show
