+++ app/controllers/home_controller.rb
@@ -1,4 +1,5 @@
 class HomeController < ApplicationController
+  include ModelFilters
   before_action :check_for_first_use
 
   def index
