+++ app/views/creators/index.html.erb
@@ -1,20 +1,19 @@
-<div class="row">
-  <div class="col">
-    <h1>Creators</h1>
-  </div>
-  <div class="col col-auto">
-    <%= link_to "View unassigned", models_path(creator: ""), class: "btn btn-outline-secondary" %>
-  </div>
-  <div class="col col-auto">
-    <%= link_to "New Creator", new_creator_path, class: "btn btn-primary" %>
-  </div>
-</div>
-
-<div>
-  <div class="row row-cols-1 row-cols-sm-2 row-cols-lg-3 row-cols-xl-4 mb-3" data-masonry='{"percentPosition": true }'>
+<%= render "application/content_header" %>
+<div class="row row-cols-2">
+  <div class="col-9">
+    <div class="row row-cols-1 row-cols-md-2 row-cols-lg-3 mb-4">
       <%= render @creators %>
+    </div>
+    <% if current_user.pagination_settings["creators"] %>
+      <%= paginate @creators %>
+    <% end %>
   </div>
-  <% if current_user.pagination_settings["creators"] %>
-    <%= paginate @creators %>
-  <% end %>
+  <div class="col-3">
+    <%= card :secondary, "Actions" do %>
+      <%= link_to "View unassigned", models_path(creator: ""), class: "btn btn-outline-secondary mb-3 me-3" %>
+      <%= link_to "New Creator", new_creator_path, class: "btn btn-primary mb-3 me-3" %>
+    <% end %>
+    <%= render 'filters_card' %>
+    <%= render 'tags_card', tags: @tags - (@tag||[]), muted_tags: @tags - @commontags %>
+  </div>
 </div>
