+++ app/views/home/index.html.erb
@@ -1,4 +1,5 @@
 <div class="row bg-light mb-5 text-center">
+  <%= render "application/content_header" %>
   <h1>VanDAM</h1>
   <p class='lead'>Helping you keep track of your 3d print files</p>
   <div class="col-8 offset-md-2">
