+++ app/views/models/show.html.erb
@@ -1,3 +1,4 @@
+<%= render "application/content_header" %>
 <h1><a data-field="model[name]" data-path="<%= library_model_path(@model.library, @model) %>" contenteditable=true><%= @model.name %></a></h1>
 <div class="row row-cols-2">
